split the dictaphone into an engine and per-orientation apps, and grow the UI, audio and touch it rests on
The dictaphone becomes light_dictaphone_core (the engine: modules, events,
Store/AudioStream seams, commands) with two interface crates over it --
light_app_dictaphone (portrait) and light_app_dictaphone_wide (landscape) --
each instantiated by its own board module on the 3.49 (dictaphone_touch349,
dictaphone_wide_touch349), one tree, its own UF2.

light-ui grows a Row layout (equal columns, min/max widths, optional
horizontal scroll) beside the existing Stack, a window that pins a column
outside a scrolling strip, and page transitions: Stack pages slide, Row pages
cover -- the incoming page rises from the bottom over the static outgoing on
entry (via a new Display::freeze_render that renders the incoming into the back
buffer for a cheap blit), and uncovers back down on return. The transition
axis comes from the canvas transform, so it is correct at any board rotation.

The dictaphone's status moves to the title bar and the recording light becomes
a flashing dot inline in it (a themable `indicator` colour, default red, in
light-ui + crush). Touch input is filtered: a tap fires only when the contact
never strays beyond the slop and rests at least TAP_MIN_HOLD_US, so a brush or
a graze is dropped -- which threads a timestamp through Ui::touch.

Audio on the 3.49 stops clicking and stops dropping out: a u64 fix for the
WAV-duration overflow that panicked on long files, the speaker muted before
the card setup rather than after, a deep playback prefetch ring, and -- the
big one -- the DAC ring refilled from a DMA-completion interrupt (the first
NVIC IRQ in the Rust port; a lock-free SPSC word ring in light-rp2, the SDK
installing the handler) so a blocked poll can no longer starve the codec. The
port keeps the polled path too, for the RAM-tight ui_demo board.

Hardware-verified on both 3.49 dictaphone apps; the ui_demo and crossfire
build unchanged.

+++ module/light_mk4_touch349/rust/src/lib.rs
@@ -814,7 +814,8 @@ impl AudioMod {
                                         return;
                                 }
                                 let data_end = file.pos().saturating_add(len).min(file.size());
-                                let ms = (data_end - file.pos()) / (u32::from(channels) * 2) * 1000 / AUDIO_SAMPLE_HZ;
+                                //   u64: frames * 1000 overflows u32 past ~4.5 min of audio
+                                let ms = ((u64::from(data_end - file.pos()) / (u64::from(channels) * 2)) * 1000 / u64::from(AUDIO_SAMPLE_HZ)) as u32;
                                 info!("play: {} -- {} ch, ~{} ms", path, channels, ms);
                                 *self.play = Some(Playback { fs, file, channels: channels as u8, data_end });
                                 return;
@@ -859,6 +860,8 @@ impl Module for AudioMod {
                 let _ = self.codec.set_adc_to_dac(false);
                 static STREAM_A: ConstStaticCell<[u32; light_rp2::i2s::STREAM_WORDS]> = ConstStaticCell::new([0; light_rp2::i2s::STREAM_WORDS]);
                 static STREAM_B: ConstStaticCell<[u32; light_rp2::i2s::STREAM_WORDS]> = ConstStaticCell::new([0; light_rp2::i2s::STREAM_WORDS]);
+                //   the POLLED path (no IRQ ring): this board is too RAM-tight for the
+                // prefetch ring and its audio is light -- beeps and tones a poll keeps fed
                 self.i2s.start_stream([STREAM_A.take(), STREAM_B.take()]);
                 self.pa.set(true);
                 info!("audio up: es8311 master at {} Hz, PIO1 mclk+dout; the UART console pins now carry audio", AUDIO_SAMPLE_HZ);
